+++ tests/testthat/test-edit.R
@@ -1,14 +1,14 @@
 context("edit")
 
 expect_user_file <- function(...) {
-  expect_true(file.exists(file.path("~", ...)))
+  expect_true(file.exists(scoped_path("user", ...)))
 }
+
 expect_user_git_file <- function(...) {
-  expect_true(file.exists(file.path(git_user_dot_home(), ...)))
+  expect_true(file.exists(scoped_git_path("user", ...)))
 }
-expect_project_file <- function(...) {
-  expect_true(file.exists(proj_path(...)))
-}
+
+expect_project_file <- function(...) expect_true(file.exists(proj_path(...)))
 
 ## testing edit_XXX("user") only on travis and appveyor, because I don't want to
 ## risk creating user-level files de novo for an actual user, which would
@@ -44,6 +44,8 @@ test_that("edit_git_XXX('user') ensures the file exists", {
 
   capture_output(edit_git_ignore("user"))
   expect_user_git_file(".gitignore")
+  cfg <- git2r::config()
+  expect_match(cfg$global$core.excludesfile, "gitignore")
 })
 
 test_that("edit_r_profile() ensures .Rprofile exists in project", {